refactor: simplify adminRoutes by creating adminController
